% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attributes.R
\name{get_attributes}
\alias{get_attributes}
\title{Get attributes from an imported SVG file}
\usage{
get_attributes(xml, node = NULL, node_attr = "label", attr = NULL)
}
\arguments{
\item{xml}{(XMLsvg) XMLsvg object obtained from `read_svg()`}

\item{node}{(character) node(s) to be displayed. Node names are searched using
the `node_attr` field}

\item{node_attr}{(character) name of the attribute by which target nodes are
filtered (default: "label")}

\item{attr}{(character) the desired attribute to be displayed, such as `id`, `style`}
}
\value{
Returns a `tibble` with rows being the selected nodes and
  columns being the selected attributes
}
\description{
SVG files and the imported representation of class `XMLsvg` are
hierarchically organized in node sets and nodes. Each node is a graphical
object such as a path, rectancle, or text. The attributes of these nodes
can be queried using this function.
}
