% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_values.R
\name{get_values}
\alias{get_values}
\title{Get values from an imported SVG file}
\usage{
get_values(xml, node = NULL, node_attr = "label")
}
\arguments{
\item{xml}{(XMLsvg) XMLsvg object obtained from \code{read_svg()}}

\item{node}{(character) node(s) to be modified. Node names are searched using
the `node_attr` field}

\item{node_attr}{(character) name of the attribute by which target nodes are
filtered (default: "label")}
}
\value{
Returns a named list with values for the queried nodes
  (or NULL if no value exists)
}
\description{
SVG files and the imported representation of class `XMLsvg` are
hierarchically organized in node sets and nodes. Each node is a graphical
object such as a path, rectangle, or text. The values of these nodes
can be queried using this function. A typical application for this
function is to query the content of a text field. If you want to retrieve
the (style) _attributes_ of a text field, use \code{get_attributes()} instead.
}
