% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metabolic_flux}
\alias{metabolic_flux}
\title{Metabolic fluxes}
\format{
A data frame with 68 rows and 3 variables
}
\usage{
data(metabolic_flux)
}
\description{
Metabolic fluxes for growth of the bacterium Cupriavidus necator on different carbon sources
}
\details{
\itemize{
  \item substrate  - carbon source
  \item reaction - central carbon metabolism reaction ID
  \item flux_mmol_gDCW_h - flux in mmol per g dry cell weight per hour, from FBA
}
}
\keyword{datasets}
