% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_svg.R
\name{read_svg}
\alias{read_svg}
\title{Read an `svg()` graphic file}
\usage{
read_svg(file, ...)
}
\arguments{
\item{file}{(character) path to `.svg` file}

\item{...}{other arguments passed down to `xmlParse`}
}
\description{
SVG (standard vector graphic) files are structured text files in XML format
(Extensible Markup Language). They contain descriptions of vector objects
such as circles, rectangles, and paths. This function 'reads' an XML file
by parsing it and stores it in an object of class `XMLsvg`.
}
\examples{

# read an svg file
# SVG <- read_svg("example.svg")

# alternatively, simply load example shipped with package
#data(example_network)
#print(rounded_rect)
#summary(example_network)

}
