% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_attributes.R
\name{set_attributes}
\alias{set_attributes}
\title{Change attributes of an imported SVG file}
\usage{
set_attributes(
  xml,
  node = NULL,
  node_attr = "label",
  attr = NULL,
  pattern = NULL,
  replacement = NULL
)
}
\arguments{
\item{xml}{(XMLsvg) XMLsvg object obtained from `read_svg()`}

\item{node}{(character) node(s) to be modified. Node names are searched using
the `node_attr` field}

\item{node_attr}{(character) name of the attribute by which target nodes are
filtered (default: "label")}

\item{attr}{(character) the desired attribute to be displayed, such as `id`, `style`}

\item{pattern}{(character) (sub-) string of the attribute vlaue to be modified}

\item{replacement}{(character) string that is used as replacement for `pattern`}
}
\value{
Returns an object of class `XMLsvg`
}
\description{
SVG files and the imported representation of class `XMLsvg` are
hierarchically organized in node sets and nodes. Each node is a graphical
object such as a path, rectangle, or text. The attributes of these nodes
can be changed with this function. Note that even when the output of
\code{set_attributes()} is not assigned to a new object, the input `XMLsvg` object
is still changed. This is because, unlike usual R work flows, attributes of
XML objects are changed on the fly (similar to python).
}
