% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_values.R
\name{set_values}
\alias{set_values}
\title{Change values of an imported SVG file}
\usage{
set_values(xml, node = NULL, node_attr = "label", value = NULL)
}
\arguments{
\item{xml}{(XMLsvg) XMLsvg object obtained from `read_svg()`}

\item{node}{(character) node(s) to be modified. Node names are searched using
the `node_attr` field}

\item{node_attr}{(character) name of the attribute by which target nodes are
filtered (default: "label")}

\item{value}{(character) The (new) value assigned to}
}
\value{
Returns an object of class `XMLsvg`
}
\description{
SVG files and the imported representation of class `XMLsvg` are
hierarchically organized in node sets and nodes. Each node is a graphical
object such as a path, rectangle, or text. This function changes _values_ of
nodes, most notably text fields. For _attributes_ which describe the properties
of a value, see \code{set_attributes()}. Note that even when the output of
`set_values()` is not assigned to a new object, the input `XMLsvg` object
is still changed. This is because, unlike usual R work flows, attributes of
XML objects are changed on the fly (similar to python).
}
