% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_svg.R
\name{write_svg}
\alias{write_svg}
\title{Write an `svg()` graphic file}
\usage{
write_svg(x, file)
}
\arguments{
\item{x}{(XMLsvg) object to export}

\item{file}{(character) File path to write to}
}
\description{
Writes an object of class `XMLsvg` to disk that was imported by `read_svg()`.
}
