% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{XMLsvg-class}
\alias{XMLsvg-class}
\alias{XMLsvg}
\title{S4 class definition for SVG XML}
\description{
SVG (standard vector grqphic) files are structured text files in XML format
(Extensible Markup Language). The new class `XMLsvg` stores the SVG file
in a list object with two slots, the SVG parsed by `parseXML()` and a summary
table.
}
